Move JoinFuzzer into exec/fuzzer to make structure clear
